Prevent use of location hash
Clicking the image gallery indicators and swiping was updating browser history to match the location hash. This caused problems when trying to go back in the browser as it became polluted with image gallery changes.

This fix prevents hashes updating the url and instead uses scrollIntoView to "manually" change the active image slide.